db_requirements.txt
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0
alembic>=1.12.0
python-dotenv>=1.0.0
